so much changes and everything is buggy, that is cool maybe; new edits on engine (and renderer); also now ui is self writted (by me, not a library) and we have now screens like in C# version

+++ client/src/engine/types/rect.v
@@ -2,8 +2,8 @@ module types
 
 pub struct Rect {
 pub mut:
-	x int
-	y int
-	w int
-	h int
+	x f32
+	y f32
+	w f32
+	h f32
 }
